+++ Tests/Integration/DSCResources/IisSite.integration.tests.ps1
@@ -19,6 +19,8 @@ try
     #region Tests
     foreach ($stig in $stigList)
     {
+        [xml] $dscXml = Get-Content -Path $stig.Path
+
         Describe "IIS Site $($stig.StigVersion) mof output" {
 
             It 'Should compile the MOF without throwing' {
@@ -31,8 +33,6 @@ try
                         -OutputPath $TestDrive
                 } | Should not throw
             }
-
-            [xml] $dscXml = Get-Content -Path $stig.Path
 
             $configurationDocumentPath = "$TestDrive\localhost.mof"
 
@@ -47,9 +47,9 @@ try
             }
 
             Context 'WebConfigurationPropertyRule' {
-               $hasAllSettings = $true
+                $hasAllSettings = $true
                 $dscXml = $dscXml.DISASTIG.WebConfigurationPropertyRule.Rule
-                $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[xWebConfigProperty\]"}
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[xWebConfigProperty\]"}
 
                 foreach ($website in $websiteName)
                 {
@@ -79,7 +79,7 @@ try
             Context 'MimeTypeRule' {
                 $hasAllSettings = $true
                 $dscXml = $dscXml.DISASTIG.MimeTypeRule.Rule
-                $dscMof = $instances | Where-Object {$PSItem.ResourceID -match "\[xIisMimeTypeMapping\]"}
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[xIisMimeTypeMapping\]"}
 
                 foreach ($website in $websiteName)
                 {
@@ -95,6 +95,82 @@ try
 
                 It "Should have $($dscXml.Count) WebConfigurationPropertyRule settings" {
                     $hasAllSettings | Should Be $true
+                }
+            }
+        }
+
+        Describe "IIS Site $($stig.StigVersion) Single SkipRule/RuleType mof output" {
+
+            $SkipRule     = Get-Random -InputObject $dscXml.DISASTIG.WebConfigurationPropertyRule.Rule.id
+            $SkipRuleType = "IisLoggingRule"
+
+            It 'Should compile the MOF without throwing' {
+                {
+                    & "$($script:DSCCompositeResourceName)_config" `
+                        -WebAppPool $WebAppPool `
+                        -WebsiteName $websiteName `
+                        -OsVersion $stig.TechnologyVersion `
+                        -StigVersion $stig.StigVersion `
+                        -OutputPath $TestDrive `
+                        -SkipRule $SkipRule `
+                        -SkipRuleType $SkipRuleType `
+                } | Should not throw
+            }
+
+            #region Gets the mof content
+            $configurationDocumentPath = "$TestDrive\localhost.mof"
+            $instances = [Microsoft.PowerShell.DesiredStateConfiguration.Internal.DscClassCache]::ImportInstances($configurationDocumentPath, 4)
+            #endregion
+
+            Context 'Skip check' {
+
+                #region counts how many Skips there are and how many there should be.
+                $dscIisLoggingRuleXml = $dscXml.DISASTIG.IisLoggingRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "pass"}
+                $expectedSkipRuleCount = ($($dscIisLoggingRuleXml.Count) + $($SkipRule.Count))
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"}
+                #endregion
+
+                It "Should have $expectedSkipRuleCount Skipped settings" {
+                    $dscMof.count | Should Be $expectedSkipRuleCount
+                }
+            }
+        }
+
+        Describe "IIS Site $($stig.StigVersion) Multiple SkipRule/RuleType mof output" {
+
+            $SkipRule     = Get-Random -InputObject $dscXml.DISASTIG.MimeTypeRule.Rule.id -Count 2
+            $SkipRuleType = @('WebAppPoolRule','IisLoggingRule')
+
+            It 'Should compile the MOF without throwing' {
+                {
+                    & "$($script:DSCCompositeResourceName)_config" `
+                        -WebsiteName $websiteName `
+                        -OsVersion $stig.TechnologyVersion `
+                        -StigVersion $stig.StigVersion `
+                        -OutputPath $TestDrive `
+                        -SkipRule $SkipRule `
+                        -SkipRuleType $SkipRuleType `
+                } | Should not throw
+            }
+
+            #region Gets the mof content
+            $configurationDocumentPath = "$TestDrive\localhost.mof"
+            $instances = [Microsoft.PowerShell.DesiredStateConfiguration.Internal.DscClassCache]::ImportInstances($configurationDocumentPath, 4)
+            #endregion
+
+            Context 'Skip check' {
+
+                #region counts how many Skips there are and how many there should be.
+                $dscWebAppPoolRuleXml = $dscXml.DISASTIG.WebAppPoolRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "Pass"}
+                $dscIisLoggingRuleXml = $dscXml.DISASTIG.IisLoggingRule.Rule | Where-Object -FilterScript {$_.ConversionStatus -eq "Pass"}
+
+                $expectedSkipRuleCount = ($($dscWebAppPoolRuleXml.Count) + $($dscIisLoggingRuleXml.count) + $($SkipRule.Count))
+
+                $dscMof = $instances | Where-Object -FilterScript {$PSItem.ResourceID -match "\[Skip\]"}
+                #endregion
+
+                It "Should have $expectedSkipRuleCount Skipped settings" {
+                    $dscMof.count | Should Be $expectedSkipRuleCount
                 }
             }
         }